feat: Provide friendlier output, don't overwrite files
The pre-existing behavior is to overwrite files and not output any logs
to say what this has done, leaving the user to hunt for what files have
changed.

Additionally, this would simply overwrite any existing files, which
could be problematic.

This change makes it such that the output now logs out `created: <file>`
for any new file created. It also skips files that already exist,
offering a `existing: <file>` log line so the user knows how to handle
this.