Allow using $EDITOR commands with arguments
When using VSCode, for the calling command to wait for the file to be
closed, you need to call `code -w`. This was failing because go
needs each part of the command to be passed in separately when using
`exec.Command`.

Using Google's `shlex` project, we can reliably parse the `$EDITOR` command
and ensure that we can call the editor properly.